move ResourceEditor to components and remove a legacy import
