+++ src/shared/components/ResourceEditor/ResourceEditor.less
