+++ src/shared/components/ResourceEditor/index.tsx
