+++ src/shared/utils/index.ts
@@ -1,5 +1,4 @@
-import { Identity } from '@bbp/nexus-sdk-legacy/lib/ACL/types';
-import { Resource, ExpandedResource } from '@bbp/nexus-sdk';
+import { Resource, Identity } from '@bbp/nexus-sdk';
 
 /**
  * getProp utility - an alternative to lodash.get
